Fix munkres build with visual studio (msvc17) and disable warnings
