+++ modules/core/test/munkres/testMunkres.cpp
@@ -47,8 +47,9 @@
 
 namespace std
 {
+
 // Helper to output a Munkres pair
-ostream &operator<<(ostream &os, const pair<uint, uint> &val)
+ostream &operator<<(ostream &os, const pair<unsigned int, unsigned int> &val)
 {
   os << "[" << val.first << "," << val.second << "]";
   return os;
@@ -62,7 +63,7 @@ ostream &operator<<(ostream &os, const pair<uint, uint> &val)
 TEST_CASE("Check Munkres-based assignment", "[visp_munkres]")
 {
   auto testMunkres = [](const std::vector<std::vector<double> > &cost_matrix,
-                        const std::vector<std::pair<uint, uint> > &expected_pairs) {
+                        const std::vector<std::pair<unsigned int, unsigned int> > &expected_pairs) {
     const auto munkres_pairs = vpMunkres::run(cost_matrix);
     REQUIRE(expected_pairs.size() == munkres_pairs.size());
     for (auto i = 0u; i < munkres_pairs.size(); i++) {
@@ -77,7 +78,7 @@ TEST_CASE("Check Munkres-based assignment", "[visp_munkres]")
     costs.push_back(std::vector<double>{2, 3, 1});
     costs.push_back(std::vector<double>{1, 2, 3});
 
-    std::vector<std::pair<uint, uint> > expected_pairs{};
+    std::vector<std::pair<unsigned int, unsigned int> > expected_pairs{};
     expected_pairs.emplace_back(0, 1);
     expected_pairs.emplace_back(1, 2);
     expected_pairs.emplace_back(2, 0);
@@ -92,7 +93,7 @@ TEST_CASE("Check Munkres-based assignment", "[visp_munkres]")
     costs.push_back(std::vector<double>{3, 4, 1, 2});
     costs.push_back(std::vector<double>{2, 3, 4, 1});
 
-    std::vector<std::pair<uint, uint> > expected_pairs{};
+    std::vector<std::pair<unsigned int, unsigned int> > expected_pairs{};
     expected_pairs.emplace_back(0, 1);
     expected_pairs.emplace_back(1, 2);
     expected_pairs.emplace_back(2, 3);
@@ -108,7 +109,7 @@ TEST_CASE("Check Munkres-based assignment", "[visp_munkres]")
     costs.push_back(std::vector<double>{2, 3, 4});
     costs.push_back(std::vector<double>{1, 2, 3});
 
-    std::vector<std::pair<uint, uint> > expected_pairs{};
+    std::vector<std::pair<unsigned int, unsigned int> > expected_pairs{};
     expected_pairs.emplace_back(0, 1);
     expected_pairs.emplace_back(1, 2);
     expected_pairs.emplace_back(3, 0);
@@ -128,7 +129,7 @@ int main(int argc, char *argv[])
 // Fallback to classic tests
 
 bool testMunkres(const std::vector<std::vector<double> > &costs_matrix,
-                 const std::vector<std::pair<uint, uint> > &expected_pairs)
+                 const std::vector<std::pair<unsigned int, unsigned int> > &expected_pairs)
 {
   const auto pairs = vpMunkres::run(costs_matrix);
 
@@ -175,7 +176,7 @@ bool testSquareMat()
   costs.push_back(std::vector<double>{1, 2, 3, 4});
   costs.push_back(std::vector<double>{2, 1, 4, 3});
 
-  std::vector<std::pair<uint, uint> > pairs{};
+  std::vector<std::pair<unsigned int, unsigned int> > pairs{};
   pairs.emplace_back(0, 2);
   pairs.emplace_back(1, 3);
   pairs.emplace_back(2, 0);
@@ -192,7 +193,7 @@ bool testVertMat()
   costs.push_back(std::vector<double>{1, 4, 3});
   costs.push_back(std::vector<double>{2, 1, 4});
 
-  std::vector<std::pair<uint, uint> > pairs{};
+  std::vector<std::pair<unsigned int, unsigned int> > pairs{};
   pairs.emplace_back(0, 2);
   pairs.emplace_back(2, 0);
   pairs.emplace_back(3, 1);
@@ -207,7 +208,7 @@ bool testHorMat()
   costs.push_back(std::vector<double>{4, 1, 2, 3});
   costs.push_back(std::vector<double>{1, 2, 3, 4});
 
-  std::vector<std::pair<uint, uint> > pairs{};
+  std::vector<std::pair<unsigned int, unsigned int> > pairs{};
   pairs.emplace_back(0, 3);
   pairs.emplace_back(1, 1);
   pairs.emplace_back(2, 0);
